pyproject.toml
[project]
name = "aoc2024"
version = "0.1.0"
description = "https://adventofcode.com/2024"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ibis-framework[datafusion,duckdb,polars] @ git+https://github.com/ibis-project/ibis.git"
]

[project.optional-dependencies]
dev = [
  "ipykernel==6.29.5",
  "ipython==8.30.0",
  "pre-commit==4.0.1",
  "pytest==8.3.4",
]

[tool.setuptools.packages]
find = { include = ["aoc2024"] }

[tool.setuptools.package-data]
"aoc2024" = ["data/**/*.txt"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
